ENH: added cast to float

+++ Projekt/Context/context.cpp
@@ -141,7 +141,7 @@ void Context::loop(std::function<void(Context&)> fixedCallback, std::function<vo
 	GLfloat lastFrame = (GLfloat) glfwGetTime();
 	double accumulator = 0.0;
 	double fixedDelta = 1 / 60.0f;
-	Context::fixedDeltaTime = fixedDelta;
+	Context::fixedDeltaTime = (GLfloat) fixedDelta;
 
 	while (!glfwWindowShouldClose(this->window))
 	{
